// enums/employee/employee-dqf.enum.ts
export enum EmployeeDqf {
  A_SECOND_FORM_OF_ID = "A_SECOND_FORM_OF_ID",
  ACCIDENT_REGISTER = "ACCIDENT_REGISTER",
  ANNUAL_MVR_REVIEW_NOTES_OPTIONAL = "ANNUAL_MVR_REVIEW_NOTES_OPTIONAL",
  BACKGROUND_CHECK_GENERAL_NON_DOT = "BACKGROUND_CHECK_GENERAL_NON_DOT",
  BANK_DEPOSIT_INFO = "BANK_DEPOSIT_INFO",
  DRIVER_AUTHORIZATION_FORMS = "DRIVER_AUTHORIZATION_FORMS",
  DRIVER_CERTIFICATION_OF_VIOLATIONS_ANNUAL = "DRIVER_CERTIFICATION_OF_VIOLATIONS_ANNUAL",
  DRIVER_LICENSE = "DRIVER_LICENSE",
  DRIVER_ROAD_TEST_CERTIFICATE_OR_EQUIVALENT = "DRIVER_ROAD_TEST_CERTIFICATE_OR_EQUIVALENT",
  DRUG_AND_ALCOHOL_CLEARINGHOUSE_QUERY_ANNUAL = "DRUG_AND_ALCOHOL_CLEARINGHOUSE_QUERY_ANNUAL",
  DRUG_AND_ALCOHOL_CUSTODY_AND_CONTROL_FORM = "DRUG_AND_ALCOHOL_CUSTODY_AND_CONTROL_FORM",
  EMERGENCY_CONTACT_LIST = "EMERGENCY_CONTACT_LIST",
  EMPLOYMENT_APPLICATION = "EMPLOYMENT_APPLICATION",
  I_9_FORM = "I_9_FORM",
  MEDICAL_EXAMINATION_REPORT_LONG_FORM = "MEDICAL_EXAMINATION_REPORT_LONG_FORM",
  MEDICAL_EXAMINER_CERTIFICATE_MEDICAL_CARD = "MEDICAL_EXAMINER_CERTIFICATE_MEDICAL_CARD",
  MOTOR_VEHICLE_RECORD_MVR = "MOTOR_VEHICLE_RECORD_MVR",
  OFFER_LETTER_AND_ACCEPTANCE = "OFFER_LETTER_AND_ACCEPTANCE",
  PRE_EMPLOYMENT_DRUG_TEST = "PRE_EMPLOYMENT_DRUG_TEST",
  PSP_RECORD = "PSP_RECORD",
  RECEIPT_FOR_COMPANY_POLICIES = "RECEIPT_FOR_COMPANY_POLICIES",
  RECEIPT_FOR_CONTROLLED_SUBSTANCE_POLICY = "RECEIPT_FOR_CONTROLLED_SUBSTANCE_POLICY",
  RECORD_OF_ATTEMPTS_SAFETY_PERFORMANCE_HISTORY = "RECORD_OF_ATTEMPTS_SAFETY_PERFORMANCE_HISTORY",
  RESUME_OPTIONAL = "RESUME_OPTIONAL",
  SAFETY_PERFORMANCE_HISTORY = "SAFETY_PERFORMANCE_HISTORY",
  SIGNED_HOURS_OF_SERVICE_RECORD_STATEMENT = "SIGNED_HOURS_OF_SERVICE_RECORD_STATEMENT",
  TRAINING_CERTIFICATION_RECORDS = "TRAINING_CERTIFICATION_RECORDS",
  TRUCK_PROVIDED = "TRUCK_PROVIDED",
  TWIC_OPTIONAL = "TWIC_OPTIONAL",
  W_4_FORM_FOR_COMPANY_DRIVER = "W_4_FORM_FOR_COMPANY_DRIVER",
  W_9_FORM_FOR_OWNER_OPERATOR = "W_9_FORM_FOR_OWNER_OPERATOR",
}